git-svn-id: http://svn.concord.org/svn/projects/trunk/common/java/simulations/mw@12253 6e01202a-0783-4428-890a-84243c50cc2b

+++ src/org/concord/modeler/HistoryManager.java
@@ -95,7 +95,7 @@ public final class HistoryManager {
 		long t1 = System.currentTimeMillis();
 		long t0;
 		for (Iterator it = map.keySet().iterator(); it.hasNext();) {
-			t0 = ((Long) map.get(it.next())).longValue();
+			t0 = map.get(it.next());
 			if (t1 - t0 > nday * DAY)
 				it.remove();
 		}
